Version 1.24
  * Made BSD keymap conversion more compatible, at the expense of true combining umlaut support.  This fixes a problem with the Brazilian keyboard layout.
  * BSD systemd unit conversion now makes use of the rtprio and idprio chain loading commands.
  * More BSD services: bsd-serial-init, static_arp@, static_ndp@, static-networking, ipfw, ipfw0, pefs, entropy, ip6addrctl@ipv4, ip6addrctl@ip6, hald, mdmfs@-tmp, mdmfs@-var
  * BSD configuration import now creates mount service bundles for ZFS volumes.
  * console-fb-realizer now attempts to pick the "best" video mode on BSD.
  * The nosh-bundles package now creates the missing /var/local/sv/ and /run/ on BSD.
  * There is now a ppp configuration import.
  * There is now an rfcomm_pppd configuration import.
  * More cross-platform services: dbmail-imapd, dbmail-pop3d
  * More Linux services: sudo

+++ source/convert/volumes.do
@@ -1,4 +1,7 @@
 #!/bin/sh -e
+## **************************************************************************
+## For copyright and licensing terms, see the file named COPYING.
+## **************************************************************************
 #
 # Convert the /etc/fstab external configuration format.
 # This is invoked by all.do .
@@ -6,13 +9,17 @@
 
 redo-ifchange /etc/fstab
 
-system-control convert-fstab-services --bundle-root /etc/service-bundles/services/ --etc-bundle --overwrite
+r="/etc/service-bundles/services/"
+o="--etc-bundle --overwrite"
+
+system-control convert-fstab-services --bundle-root "$r" $o
 
 for i in mount fsck swap dump
 do
-	for j in /etc/service-bundles/services/$i@*
+	for j in "$r$i@"*
 	do
-		n="${j#/etc/service-bundles/services/$i@}"
+		test -e "$j/" || continue
+		n="${j#$r$i@}"
 		rm -f -- "$j/log"
 		ln -s -- "../sysinit-log" "$j/log"
 		system-control disable "$i@$n"
@@ -28,9 +35,28 @@ done
 
 case "`uname`" in
 *BSD)
+	# ZFS non-swap volumes
+	zfs list -H -o canmount,name,mountpoint -t filesystem |
+	awk "{
+		if (\$3 == \"none\" || \$1 == \"no\") next;
+		print \"system-control write-volume-service-bundles --bundle-root \\\"$r\\\" $o zfs \\\"\"\$2\"\\\" \\\"\"\$3\"\\\"\";
+	}" | 
+	sh -e
+	zfs list -H -o canmount,mountpoint -t filesystem |
+	awk '{ if ($2 != "none" && $1 == "noauto") { gsub("-","\\x2d",$2); gsub("/","-",$2); print "system-control disable mount@\""$2"\""; } }' | 
+	sh -e
+	zfs list -H -o canmount,mountpoint -t filesystem |
+	awk '{ if ($2 != "none" && $1 == "on") { gsub("-","\\x2d",$2); gsub("/","-",$2); print "system-control enable mount@\""$2"\""; } }' | 
+	sh -e
+	zfs list -H -o canmount,mountpoint -t filesystem |
+	awk '{ if ($2 != "none") { gsub("-","\\x2d",$2); gsub("/","-",$2); print $1" mount@"$2; } }' >> "$3"
+	zfs list -H -o canmount,mountpoint -t filesystem / |
+	awk '{ if ($2 != "none" && $1 != "off") { gsub("-","\\x2d",$2); gsub("/","-",$2); print "system-control enable mount@\""$2"\""; } }' | 
+	sh -e
+
 	# ZFS volumes can have swap enabled via a property org.freebsd:swap=on .
 	zfs list -H -o org.freebsd:swap,name -t volume |
-	while read state n
+	while read -r state n
 	do
 		case "${state}" in
 		[Oo][Nn])
@@ -51,4 +77,4 @@ case "`uname`" in
 	;;
 esac
 
-redo-ifchange geom
+redo-ifchange geom mdmfs
